requirements.txt
fastapi
uvicorn
pydantic
openslide-python
Pillow
numpy
instanseg